+++ AddOns/AllTheThings/src/Classes/base.lua
@@ -224,6 +224,11 @@ else
 	end
 end
 
+local function ClassError(...)
+	local params = {...}
+	local err = app.TableConcat(params, nil, "", " ")
+	error(err)
+end
 local CloneDictionary = app.CloneDictionary
 -- Creates a Base Object Table which will evaluate the provided set of 'fields' (each field value being a keyed function)
 local classDefinitions, _cache = {}, nil;
@@ -234,13 +239,13 @@ end
 -- Generates a metatable to use for the given class name based on the provided field functions
 local CreateClassMeta = not app.__perf and function(fields, className)
 	if not className then
-		print("A Class Name must be declared when using CreateClass");
+		ClassError("A Class Name must be declared when using CreateClass");
 	end
 	local class = { __type = function() return className; end };
 	if not classDefinitions[className] then
 		classDefinitions[className] = class;
 	else
-		print("A Class has already been defined with that name!", className);
+		ClassError("A Class has already been defined with that name!", className);
 	end
 
 	if fields then CloneDictionary(fields, class) end
@@ -260,7 +265,7 @@ end
 -- performance tracking wrapped base fields
 or function(fields, className)
 	if not className then
-		print("A Class Name must be declared when using CreateClassMeta");
+		ClassError("A Class Name must be declared when using CreateClassMeta");
 	end
 	local class = { __type = function() return className; end };
 	app.__perf.AutoCaptureTable(class, "Class:"..className)
@@ -270,7 +275,7 @@ or function(fields, className)
 	if not classDefinitions[className] then
 		classDefinitions[className] = class;
 	else
-		print("A Class has already been defined with that name!", className);
+		ClassError("A Class has already been defined with that name!", className);
 	end
 
 	if fields then CloneDictionary(fields, class) end
@@ -404,6 +409,48 @@ app.CreateClassInstance = CreateClassInstance;
 -- wouldn't even be used obviously... so maybe in the future
 local GlobalVariants = {}
 app.GlobalVariants = GlobalVariants
+-- Takes multiple other variants and combines them together into the specified name, then storing it
+-- into the GlobalVariants table
+GlobalVariants.Combine = function(...)
+	local combine, conditions, name = {}, {}, ""
+	local condition, variantName
+	-- combine tables, check unique fields
+	for _,variant in ipairs({...}) do
+		variantName = variant.__name
+		if not variantName or type(variantName) ~= "string" then
+			ClassError("Cannot combine variants due to variant",variantName or _,"missing valid '__name' string!")
+		end
+		name = name..variantName
+		condition = variant.__condition
+		if not condition or type(condition) ~= "function" then
+			ClassError("Cannot combine variants due to variant",variantName,"missing '__condition' function!")
+		end
+		conditions[#conditions + 1] = condition
+		for k, v in pairs(variant) do
+			if k ~= "__condition" and k ~= "__name" then
+				if combine[k] then
+					ClassError("Cannot combine variants due to conflicting variant key:",k)
+				end
+				combine[k] = v
+			end
+		end
+	end
+	-- save into GlobalVariants
+	if GlobalVariants[name] then
+		ClassError("Cannot Combine variants into",name,"since it has already been defined!")
+	end
+	-- store the combined name
+	combine.__name = name
+	-- create a new __condition based on the running of all other conditions
+	combine.__condition = function(t)
+		for _,condition in ipairs(conditions) do
+			if not condition(t) then return end
+		end
+		return true
+	end
+	GlobalVariants[name] = combine
+	return combine
+end
 
 local function GenerateVariantClasses(class)
 	local fields = class.__class
@@ -412,14 +459,16 @@ local function GenerateVariantClasses(class)
 	local subbase = function(t, key) return class.__index; end
 	local classname = fields.__type()
 	local variantClone
-	for variantName,variant in pairs(variants) do
+	for i,variant in ipairs(variants) do
+		if not variant.__name then
+			ClassError("Missing Class Variant __name!",i,classname)
+		end
 		if not variant.__condition then
-			app.print("Missing Class Variant __condition!",variantName,classname)
-			error("ATT Variants Require __condition function")
+			ClassError("Missing Class Variant __condition!",variant.__name,classname)
 		end
 		-- raw variant table may be used by other classes, so need to copy it for this specific subclass
 		variantClone = CloneDictionary(fields, CloneDictionary(variant, {base=subbase}))
-		variants[variantName] = CreateClassMeta(variantClone, classname..variantName);
+		variants[i] = CreateClassMeta(variantClone, classname..variant.__name);
 	end
 end
 local function AppendVariantConditionals(conditionals, class)
@@ -430,10 +479,10 @@ local function AppendVariantConditionals(conditionals, class)
 			conditionals[#conditionals + 1] = function(t)
 				if subcassCondition(t) then
 					-- check any variants for this subclass
-					for variantName,variant in pairs(variants) do
+					for i,variant in ipairs(variants) do
 						if variant.__class.__condition(t) then
 							setmetatable(t, variant);
-							-- app.PrintDebug("Create Variant",t.hash,class.__class.__type()..variantName)
+							-- app.PrintDebug("Create Variant",t.hash,class.__class.__type()..variant.__name)
 							return true
 						end
 					end
@@ -452,10 +501,10 @@ local function AppendVariantConditionals(conditionals, class)
 	elseif variants then
 		conditionals[#conditionals + 1] = function(t)
 			-- check any variants for this class
-			for variantName,variant in pairs(variants) do
+			for i,variant in ipairs(variants) do
 				if variant.__class.__condition(t) then
 					setmetatable(t, variant);
-					-- app.PrintDebug("Create Variant",t.hash,class.__class.__type()..variantName)
+					-- app.PrintDebug("Create Variant",t.hash,class.__class.__type()..variant.__name)
 					return true
 				end
 			end
@@ -482,13 +531,13 @@ end
 app.CreateClass = function(className, classKey, fields, ...)
 	-- Validate arguments
 	if not className then
-		print("A Class Name must be declared when using CreateClass");
+		ClassError("A Class Name must be declared when using CreateClass");
 	end
 	if not classKey then
-		print("A Class Key must be declared when using CreateClass");
+		ClassError("A Class Key must be declared when using CreateClass");
 	end
 	if not fields then
-		print("Fields must be declared when using CreateClass");
+		ClassError("Fields must be declared when using CreateClass");
 	end
 
 	-- app.PrintDebug("CreateClass",className, classKey)
@@ -544,7 +593,7 @@ app.CreateClass = function(className, classKey, fields, ...)
 	if not classesByKey[classKey] then
 		classesByKey[classKey] = classConstructor;
 	elseif not fields.IsClassIsolated then
-		print(className, "does not have a unique class Key", classKey, "and will have trouble with instance creation without a direct reference to an existing object or a direct integration using parser!");
+		ClassError(className, "does not have a unique class Key", classKey, "and will have trouble with instance creation without a direct reference to an existing object or a direct integration using parser!");
 	end
 	return classConstructor, Class;
 end
@@ -554,16 +603,16 @@ end
 app.CreateClassWithInfo = function(className, classKey, classInfo, fields)
 	-- Validate arguments
 	if not className then
-		print("A Class Name must be declared when using CreateClassWithInfo");
+		ClassError("A Class Name must be declared when using CreateClassWithInfo");
 	end
 	if not classKey then
-		print("A Class Key must be declared when using CreateClassWithInfo");
+		ClassError("A Class Key must be declared when using CreateClassWithInfo");
 	end
 	if not fields then
-		print("Fields must be declared when using CreateClassWithInfo");
+		ClassError("Fields must be declared when using CreateClassWithInfo");
 	end
 	if not classInfo then
-		print("ClassInfo must be declared when using CreateClassWithInfo");
+		ClassError("ClassInfo must be declared when using CreateClassWithInfo");
 	end
 
 	-- Ensure that a key and _type field exists!
@@ -574,7 +623,7 @@ app.CreateClassWithInfo = function(className, classKey, classInfo, fields)
 	if not classDefinitions[className] then
 		classDefinitions[className] = class;
 	else
-		print("A Class has already been defined with that name!", className);
+		ClassError("A Class has already been defined with that name!", className);
 	end
 
 	local classConstructor = function(id, t)
@@ -622,7 +671,7 @@ app.ExtendClass = function(baseClassName, className, classKey, fields, ...)
 		fields.conditionals = nil;
 		fields.simplemeta = nil;
 	else
-		print("Could not find specified base class:", baseClassName);
+		ClassError("Could not find specified base class:", baseClassName);
 	end
 	-- app.PrintDebug("ExtendClass",baseClassName, className, classKey)
 	return app.CreateClass(className, classKey, fields, ...);
@@ -865,7 +914,7 @@ local fieldsWithWorldQuest = {
 		end
 	end,
 	variants = {
-		AndReputation = {
+		{
 			Test = function(t)
 				return "AndReputation";
 			end,
@@ -874,6 +923,7 @@ local fieldsWithWorldQuest = {
 					print(t.name .. " (" .. t.__type .. "): I'm a subvariant with reputation!");
 				end
 			end,
+			__name = "VariantName",
 			__condition = function(t)
 				return t.maxReputation
 			end,