+++ AddOns/Atlas/Core/AtlasFrame.lua
@@ -30,7 +30,7 @@ local WoWClassicEra, WoWClassic, WoWRetail
 local wowversion = select(4, GetBuildInfo())
 if wowversion < 20000 then
 	WoWClassicEra = true
-elseif wowversion > 30000 and wowversion < 90000 then
+elseif wowversion > 20000 and wowversion < 90000 then
 	WoWClassic = true
 elseif wowversion > 90000 then
 	WoWRetail = true