+++ AddOns/Atlas/Integration/Achievements.lua
@@ -24,41 +24,35 @@
 --]]
 
 -- Determine WoW TOC Version
-local WoWClassicEra, WoWClassicTBC, WoWWOTLKC, WoWRetail
+local WoWClassicEra, WoWClassic, WoWRetail
 local wowversion = select(4, GetBuildInfo())
 if wowversion < 20000 then
 	WoWClassicEra = true
-elseif wowversion < 30000 then
-	WoWClassicTBC = true
-elseif wowversion < 40000 then
-	WoWWOTLKC = true
+elseif wowversion > 20000 and wowversion < 90000 then
+	WoWClassic = true
 elseif wowversion > 90000 then
 	WoWRetail = true
 end
 
--- ----------------------------------------------------------------------------
--- AddOn namespace.
--- ----------------------------------------------------------------------------
 local FOLDER_NAME, private = ...
-local LibStub = _G.LibStub
 local addon = LibStub("AceAddon-3.0"):GetAddon(private.addon_name)
 local L = LibStub("AceLocale-3.0"):GetLocale(private.addon_name)
 
 function addon:AchievementButtonUpdate(button, achievementID)
-	-- WOLTKC starts to introduce achievement system, so we are only skipping this for Classic Era and TBC
-	if (WoWClassicEra or WoWClassicTBC) then return end
+	-- WOLTKC starts to introduce achievement system, so we are only skipping this for Classic Era
+	if (WoWClassicEra) then return end
 
 	button.achievementID = achievementID
 	button.link = GetAchievementLink(achievementID) or nil
-	-- id, name, points, completed, month, day, year, description, flags, icon, rewardText, isGuild, wasEarnedByMe, earnedBy = GetAchievementInfo(achievementID or categoryID, index)
+
 	local _, name, _, completed, month, day, year, description, _, icon, _, _, _, earnedBy = GetAchievementInfo(achievementID)
 	if (name) then
 		button.tooltiptitle = format("|T%d:0:0|t |cFFFFFFFF%s|r", icon, name)
 		local tooltiptext = description
 		local numCriteria = GetAchievementNumCriteria(achievementID)
+
 		if (numCriteria and numCriteria > 0) then
 			for i = 1, numCriteria do
-				-- criteriaString, criteriaType, completed, quantity, reqQuantity, charName, flags, assetID, quantityString, criteriaID, eligible =  GetAchievementCriteriaInfo(achievementID, criteriaIndex)
 				local criteriaString, criteriaType, criteriaCompleted, quantity, reqQuantity, _, flags, assetID, quantityString = GetAchievementCriteriaInfo(achievementID, i)
 				if (criteriaType) then
 					if (criteriaType == CRITERIA_TYPE_ACHIEVEMENT and assetID) then
@@ -68,8 +62,8 @@ function addon:AchievementButtonUpdate(button, achievementID)
 						else
 							tooltiptext = tooltiptext.."\n|CFF808080 - "..aname
 						end
-						--elseif (criteriaString == "" or reqQuantity > 1) then
 					elseif (bit.band(flags, EVALUATION_TREE_FLAG_PROGRESS_BAR) == EVALUATION_TREE_FLAG_PROGRESS_BAR) then
+						--elseif (criteriaString == "" or reqQuantity > 1) then
 						if (quantity >= reqQuantity) then
 							tooltiptext = tooltiptext.."\n|CFFFFFFFF - "..quantityString
 						else
@@ -85,26 +79,30 @@ function addon:AchievementButtonUpdate(button, achievementID)
 				end
 			end
 		end
+
 		if (completed) then
 			name = format("      |T%d:0:0|t |CFFFFFFFF%s", icon, name)
 			tooltiptext = tooltiptext.."\n|CFF00FF00"..format(ACHIEVEMENT_TOOLTIP_COMPLETE, earnedBy, month, day, year)
 		else
 			name = format("      |T%d:0:0|t |CFF808080%s", icon, name)
 		end
+
 		button.Text:SetText(name)
-		button.tooltiptext = tooltiptext --.."\n|CFF8080FF"..L["ATLAS_OPEN_ACHIEVEMENT"].."|R"
+		button.tooltiptext = tooltiptext.."\n|CFF8080FF"..L["ATLAS_OPEN_ACHIEVEMENT"].."|R"
 	end
 end
 
 function addon:OpenAchievement(achievementID)
-	-- WOLTKC starts to introduce achievement system, so we are only skipping this for Classic Era and TBC
-	if (WoWClassicEra or WoWClassicTBC) then return end
+	-- WOLTKC starts to introduce achievement system, so we are only skipping this for Classic Era
+	if (WoWClassicEra) then return end
 
 	if not achievementID then return end
 
 	if not C_AddOns.IsAddOnLoaded("Blizzard_AchievementUI") then
 		C_AddOns.LoadAddOn("Blizzard_AchievementUI")
 	end
-	--ShowUIPanel(AchievementFrame)
-	--AchievementFrame_SelectAchievement(achievementID, true)
+
+	HideUIPanel(AchievementFrame)
+	ShowUIPanel(AchievementFrame)
+	AchievementFrame_SelectAchievement(achievementID, true)
 end