+++ AddOns/DBM-Raids-Vanilla/BlackrockDepths/EmperorDagranThaurissan.lua
@@ -3,11 +3,11 @@ local L		= mod:GetLocalizedStrings()
 
 mod.statTypes = "lfr,normal,heroic"
 
-mod:SetRevision("20241028071821")
+mod:SetRevision("20241029104837")
 mod:SetCreatureID(226305, 226314)
 mod:SetEncounterID(3049)
 --mod:SetUsedIcons(8, 7, 6)
---mod:SetHotfixNoticeRev(20220322000000)
+mod:SetHotfixNoticeRev(20241028000000)
 --mod:SetMinSyncRevision(20211203000000)
 --mod.respawnTime = 29