+++ AddOns/MythicDungeonTools/Modules/Transmission.lua
@@ -354,7 +354,11 @@ hooksecurefunc("SetItemRef", function(link, text)
     local sender = link:sub(17, string.len(link))
     local name, realm = string.match(sender, "(.*)+(.*)")
     if (not name) or (not realm) then
-      print(string.format(L["receiveErrorUpdate"], sender))
+      local msg = "\nsender: "..sender
+      local escapedText = text:gsub("|", "||")
+      msg = msg.."\nfull text: "..escapedText
+      local cache = MDT.U.TableToString(MDT.transmissionCache)
+      MDT:OnError(msg, cache, "MDT failed to import preset from chat link")
       return
     end
     -- to get the displayName (name of the preset) we need to get everything between the starting and closing brackets
@@ -367,7 +371,12 @@ hooksecurefunc("SetItemRef", function(link, text)
         MDT:ImportPreset(CopyTable(preset))
       end, "showInterfaceChatImport")
     else
-      print(string.format(L["receiveErrorUpdate"], sender))
+      local msg = "\nparsed displayName: "..displayName
+      msg = msg.."\nsender: "..sender
+      local escapedText = text:gsub("|", "||")
+      msg = msg.."\nfull text: "..escapedText
+      local cache = MDT.U.TableToString(MDT.transmissionCache)
+      MDT:OnError(msg, cache, "MDT failed to import preset from chat link")
     end
     return
   end
@@ -393,7 +402,7 @@ function MDTcommsObject:OnCommReceived(prefix, message, distribution, sender)
   --the user still decides if he wants to click the chat link and add the preset to his db
   if prefix == presetCommPrefix then
     local preset = MDT:StringToTable(message, false)
-    local dungeon = MDT:GetDungeonName(preset.value.currentDungeonIdx)
+    local dungeon = MDT:GetDungeonName(preset.value.currentDungeonIdx, true)
     local presetName = preset.text
     local displayName = dungeon..": "..presetName
     MDT.transmissionCache[fullName] = MDT.transmissionCache[fullName] or {}
@@ -653,7 +662,8 @@ function MDTcommsObject:OnCommReceived(prefix, message, distribution, sender)
   if prefix == MDT.liveSessionPrefixes.preset then
     if MDT.liveSessionActive then
       local preset = MDT:StringToTable(message, false)
-      local displayName = MDT:GetDungeonName(preset.value.currentDungeonIdx)..": "..preset.text
+      local dungeon = MDT:GetDungeonName(preset.value.currentDungeonIdx, true)
+      local displayName = dungeon..": "..preset.text
       MDT.transmissionCache[fullName] = MDT.transmissionCache[fullName] or {}
       MDT.transmissionCache[fullName][displayName] = preset
       if MDT:ValidateImportPreset(preset) then
@@ -728,7 +738,7 @@ local function displaySendingProgress(userArgs, bytesSent, bytesToSend)
     --output chat link
     if not silent and preset then
       local prefix = "[MDT_v2: "
-      local dungeon = MDT:GetDungeonName(preset.value.currentDungeonIdx)
+      local dungeon = MDT:GetDungeonName(preset.value.currentDungeonIdx, true)
       local presetName = preset.text
       local name, realm = UnitFullName("player")