+++ AddOns/RaidAchievement_OldModules/RaidAchievement_OldModules.lua
@@ -1,4 +1,4 @@
 ﻿function RaidAchievement_OldModules_OnLoad()
-  raOldModVers=11.004
+  raOldModVers=11.005
   raoldmodules=1
 end