+++ AddOns/RaidAchievement_PandaHeroics/RaidAchievement_PandaHeroics.lua
@@ -5,7 +5,7 @@ function phraonload()
 	_, phraenglishclass = UnitClass("player")
 	phrahuntertime=0
 	
-	raPandaModVers=11.005
+	raPandaModVers=11.006
 
 local _, a2 = GetInstanceInfo()
 if select(3,GetInstanceInfo())==2 and a2=="party" then