+++ AddOns/TradeSkillMaster/ChangeLog.md
@@ -1,6 +1,8 @@
-## v4.14.19 Changes
+## v4.14.20 Changes
 
-* [Retail] Removed restocking of crafts which require concentration
-* [Retail] Added concentration to the detailed crafting cost tooltip
+* [Retail] Updated for 11.0.5
+* [Classic] Updated for 4.4.1
+* [Classic] Crafting related fixes for non-english clients
+* Fix for clearing unread mails
 
 [Known Issues](https://support.tradeskillmaster.com/en_US/known_issues)